+++ catscripts/Entertainment/Movies/The_Lion_King/TheLionKing.sh
@@ -2,15 +2,23 @@
 
 export KEYWORDS_MOVIES_THELIONKING="Lion(| )King|Simba|Mufasa|\bNala|Sarabi|Sarafina|Shenzi|Rafiki|\bTimon|Pumbaa|Kiara|Kovu|\bZira|Vitani|Pride Rock|Pridelands"
 
-egrep -i "$KEYWORDS_MOVIES_THELIONKING" newpages.txt >> TheLionKing.txt
-
-LIONKING=$(stat --print=%s TheLionKing.txt)
-
-if [ $LIONKING -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="TheLionKing.txt"
-  export CATNAME="The Lion King"
-  $CATEGORIZE
-fi
 
-rm TheLionKing.txt
+  debug_start "The Lion King"
+
+  LIONKING=$(egrep -i "$KEYWORDS_MOVIES_THELIONKING" newpages.txt)
+
+  if [ "$LIONKING" != "" ];
+  then
+    printf "%s" "$LIONKING" > TheLionKing.txt
+    export CATFILE="TheLionKing.txt"
+    export CATNAME="The Lion King"
+    $CATEGORIZE
+    rm TheLionKing.txt
+    unset LIONKING
+  fi
+
+  debug_end "The Lion King"
+
+fi
