+++ catscripts/Entertainment/Video_games/Video_game_consoles/Nintendo/NintendoWii.sh
@@ -3,15 +3,23 @@
 KEYWORDS_NINTENDO_WII="\bWii"
 KEYWORDS_NINTENDO_WII_EXCLUDE="\bWii(| )U\b|\bWii(| )Sports|\bWii(| )Fit|\bWii(| )Play|Mario(| )Kart(| )Wii"
 
-egrep -i "$KEYWORDS_NINTENDO_WII" newpages.txt | egrep -iv "$KEYWORDS_NINTENDO_WII_EXCLUDE" >> NintendoWii.txt
-
-WII=$(stat --print=%s NintendoWii.txt)
-
-if [ $WII -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="NintendoWii.txt"
-  export CATNAME="Nintendo Wii"
-  $CATEGORIZE
-fi
 
-rm NintendoWii.txt
+  debug_start "Nintendo Wii"
+
+  WII=$(egrep -i "$KEYWORDS_NINTENDO_WII" newpages.txt | egrep -iv "$KEYWORDS_NINTENDO_WII_EXCLUDE")
+
+  if [ "$WII" != "" ];
+  then
+    printf "%s" "$WII" > NintendoWii.txt
+    export CATFILE="NintendoWii.txt"
+    export CATNAME="Nintendo Wii"
+    $CATEGORIZE
+    rm NintendoWii.txt
+    unset WII
+  fi
+
+  debug_end "Nintendo Wii"
+
+fi
