+++ catscripts/Government/Countries/Northern_Ireland/NorthernIreland.sh
@@ -1,24 +1,37 @@
 #!/bin/bash
 
-egrep -i 'Northern(| )Ireland' newpages.txt | egrep -iv 'Belfast' >> NorthernIreland.txt
-egrep -i 'Belfast' newpages.txt >> Belfast.txt
+KEYWORDS_NORTHERNIRELAND="Northern(| )Ireland"
+KEYWORDS_BELFAST="Belfast"
+KEYWORDS_NORTHERNIRELAND_EXCLUDE="$KEYWORDS_BELFAST"
 
-NIRELAND=$(stat --print=%s NorthernIreland.txt)
-BELFAST=$(stat --print=%s Belfast.txt)
-
-if [ $NIRELAND -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="NorthernIreland.txt"
-  export CATNAME="Northern Ireland"
-  $CATEGORIZE
-fi
 
-if [ $BELFAST -ne 0 ];
-then
-  export CATFILE="Belfast.txt"
-  export CATNAME="Belfast"
-  $CATEGORIZE
-fi
+  debug_start "Northern Ireland"
 
-rm NorthernIreland.txt
-rm Belfast.txt
+  NIRELAND=$(egrep -i "$KEYWORDS_NORTHERNIRELAND" newpages.txt | egrep -iv "$KEYWORDS_NORTHERNIRELAND_EXCLUDE")
+  BELFAST=$(egrep -i "$KEYWORDS_BELFAST" newpages.txt)
+
+  if [ "$NIRELAND" != "" ];
+  then
+    printf "%s" "$NIRELAND" > NorthernIreland.txt
+    export CATFILE="NorthernIreland.txt"
+    export CATNAME="Northern Ireland"
+    $CATEGORIZE
+    rm NorthernIreland.txt
+    unset NIRELAND
+  fi
+
+  if [ "$BELFAST" != "" ];
+  then
+    printf "%s" "$BELFAST" > Belfast.txt
+    export CATFILE="Belfast.txt"
+    export CATNAME="Belfast"
+    $CATEGORIZE
+    rm Belfast.txt
+    unset BELFAST
+  fi
+
+  debug_end "Northern Ireland"
+
+fi
