+++ catscripts/Government/Countries/United_States/US_states/Georgia/Georgia.sh
@@ -5,25 +5,34 @@ KEYWORDS_ATLANTA_EXCLUDE="Atlanta(| )Braves"
 KEYWORDS_GEORGIA="Georgia|, GA\b"
 KEYWORDS_GEORGIA_EXCLUDE="Republic(| )of(| )Georgia|$KEYWORDS_ATLANTA"
 
-egrep -i "$KEYWORDS_GEORGIA" newpages.txt | egrep -iv "$KEYWORDS_GEORGIA_EXCLUDE" >> Georgia.txt
-egrep -i "$KEYWORDS_ATLANTA" newpages.txt | egrep -iv "$KEYWORDS_ATLANTA_EXCLUDE" >> Atlanta.txt
-
-GEORGIA=$(stat --print=%s Georgia.txt)
-ATLANTA=$(stat --print=%s Atlanta.txt)
-
-if [ $GEORGIA -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="Georgia.txt"
-  export CATNAME="Georgia"
-  $CATEGORIZE
-fi
 
-if [ $ATLANTA -ne 0 ];
-then
-  export CATFILE="Atlanta.txt"
-  export CATNAME="Atlanta"
-  $CATEGORIZE
-fi
+  debug_start "Georgia"
 
-rm Georgia.txt
-rm Atlanta.txt
+  GEORGIA=$(egrep -i "$KEYWORDS_GEORGIA" newpages.txt | egrep -iv "$KEYWORDS_GEORGIA_EXCLUDE")
+  ATLANTA=$(egrep -i "$KEYWORDS_ATLANTA" newpages.txt | egrep -iv "$KEYWORDS_ATLANTA_EXCLUDE")
+
+  if [ "$GEORGIA" != "" ];
+  then
+    printf "%s" "$GEORGIA" > Georgia.txt
+    export CATFILE="Georgia.txt"
+    export CATNAME="Georgia"
+    $CATEGORIZE
+    rm Georgia.txt
+    unset GEORGIA
+  fi
+
+  if [ "$ATLANTA" != "" ];
+  then
+    printf "%s" "$ATLANTA" > Atlanta.txt
+    export CATFILE="Atlanta.txt"
+    export CATNAME="Atlanta"
+    $CATEGORIZE
+    rm Atlanta.txt
+    unset ATLANTA
+  fi
+
+  debug_end "Georgia"
+
+fi
