+++ catscripts/Government/Countries/Wales/Wales.sh
@@ -4,25 +4,34 @@ KEYWORDS_WALES="Wales"
 KEYWORDS_CARDIFF="Cardiff"
 KEYWORDS_WALES_EXCLUDE="$KEYWORDS_CARDIFF|Jim(|bo|my)(| )Wales|New(| )South(| )Wales"
 
-egrep -i "$KEYWORDS_WALES" newpages.txt | egrep -iv "$KEYWORDS_WALES_EXCLUDE" >> Wales.txt
-egrep -i "$KEYWORDS_CARDIFF" newpages.txt >> Cardiff.txt
-
-WALES=$(stat --print=%s Wales.txt)
-CARDIFF=$(stat --print=%s Cardiff.txt)
-
-if [ $WALES -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="Wales.txt"
-  export CATNAME="Wales"
-  $CATEGORIZE
-fi
 
-if [ $CARDIFF -ne 0 ];
-then
-  export CATFILE="Cardiff.txt"
-  export CATNAME="Cardiff"
-  $CATEGORIZE
-fi
+  debug_start "Wales"
 
-rm Wales.txt
-rm Cardiff.txt
+  WALES=$(egrep -i "$KEYWORDS_WALES" newpages.txt | egrep -iv "$KEYWORDS_WALES_EXCLUDE")
+  CARDIFF=$(egrep -i "$KEYWORDS_CARDIFF" newpages.txt)
+
+  if [ "$WALES" != "" ];
+  then
+    printf "%s" "$WALES" > Wales.txt
+    export CATFILE="Wales.txt"
+    export CATNAME="Wales"
+    $CATEGORIZE
+    rm Wales.txt
+    unset WALES
+  fi
+
+  if [ "$CARDIFF" != "" ];
+  then
+    printf "%s" "$CARDIFF" > Cardiff.txt
+    export CATFILE="Cardiff.txt"
+    export CATNAME="Cardiff"
+    $CATEGORIZE
+    rm Cardiff.txt
+    unset CARDIFF
+  fi
+
+  debug_end "Wales"
+
+fi
