+++ catscripts/Lifestyle/Health/Diseases/Infectious_diseases/Flu/Flu.sh
@@ -1,24 +1,37 @@
 #!/bin/bash
 
-egrep -i 'Flu\b|influenza' newpages.txt | egrep -iv 'H1N1|swine flu' >> Flu.txt
-egrep -i 'H1N1|Swine flu' newpages.txt >> SwineFlu.txt
+KEYWORDS_FLU="\bFlu\b|influenza"
+KEYWORDS_SWINEFLU="H1N1|swine(| )flu"
+KEYWORDS_FLU_EXCLUDE="$KEYWORDS_SWINEFLU"
 
-FLU=$(stat --print=%s Flu.txt)
-SWINE=$(stat --print=%s SwineFlu.txt)
-
-if [ $FLU -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="Flu.txt"
-  export CATNAME="Flu"
-  $CATEGORIZE
-fi
 
-if [ $SWINE -ne 0 ];
-then
-  export CATFILE="SwineFlu.txt"
-  export CATNAME="Swine flu"
-  $CATEGORIZE
-fi
+  debug_start "Flu"
 
-rm Flu.txt
-rm SwineFlu.txt
+  FLU=$(egrep -i "$KEYWORDS_FLU" newpages.txt | egrep -iv "$KEYWORDS_FLU_EXCLUDE")
+  SWINE=$(egrep -i "$KEYWORDS_SWINEFLU" newpages.txt)
+
+  if [ "$FLU" != "" ];
+  then
+    printf "%s" "$FLU" > Flu.txt
+    export CATFILE="Flu.txt"
+    export CATNAME="Flu"
+    $CATEGORIZE
+    rm Flu.txt
+    unset FLU
+  fi
+
+  if [ "$SWINE" != "" ];
+  then
+    printf "%s" "$SWINE" > SwineFlu.txt
+    export CATFILE="SwineFlu.txt"
+    export CATNAME="Swine flu"
+    $CATEGORIZE
+    rm SwineFlu.txt
+    unset SWINE
+  fi
+
+  debug_end "Flu"
+
+fi
