+++ catscripts/Math/Multiplication/Multiplication.sh
@@ -1,16 +1,24 @@
 #!/bin/bash
 
-egrep -i '[0-9]{1,} times [0-9]{1,}' newpages.txt >> Multiplication.txt
-egrep -i 'What is [0-9]{1,} x [0-9]{1,}' newpages.txt >> Multiplication.txt
-egrep -i '[0-9]{1,} \* [0-9]{1,}' newpages.txt >> Multiplication.txt
+KEYWORDS_MULTIPLICATION="[0-9]{1,}(| )times(| )[0-9]{1,}|What(| )is(| )[0-9]{1,}(| )x(| )[0-9]{1,}|[0-9]{1,}(| )\*(| )[0-9]{1,}"
 
-MULTIPLICATION=$(stat --print=%s Multiplication.txt)
-
-if [ $MULTIPLICATION -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="Multiplication.txt"
-  export CATNAME="Multiplication"
-  $CATEGORIZE
-fi
 
-rm Multiplication.txt
+  debug_start "Multiplication"
+
+  MULTIPLICATION=$(egrep -i "$KEYWORDS_MULTIPLICATION" newpages.txt)
+
+  if [ "$MULTIPLICATION" != "" ];
+  then
+    printf "%s" "$MULTIPLICATION" > Multiplication.txt
+    export CATFILE="Multiplication.txt"
+    export CATNAME="Multiplication"
+    $CATEGORIZE
+    rm Multiplication.txt
+    unset MULTIPLICATION
+  fi
+
+  debug_end "Multiplication"
+
+fi
